% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{call_vision_api}
\alias{call_vision_api}
\title{helper function to send POST request to the Google Vision API}
\usage{
call_vision_api(
  body,
  apiEndpoint = "images:annotate",
  httpRequestType = "POST"
)
}
\arguments{
\item{body, }{output of create_request_body()}

\item{apiEndpoint}{character, api endpoint}

\item{httpRequestType}{character, type of the http request}
}
\value{
API response in raw format
}
\description{
sends the request defined in `body` to the API
}
