% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{create_request_body}
\alias{create_request_body}
\title{helper function to create json for response request}
\usage{
create_request_body(txt, feature, maxNumResults)
}
\arguments{
\item{txt, }{output of image_to_text()}

\item{feature}{string, one out of: "FACE_DETECTION", "LANDMARK_DETECTION",
"LOGO_DETECTION", "LABEL_DETECTION", "TEXT_DETECTION"}

\item{maxNumResults}{numeric, the maximnum number of rows returned by the
Google Vision API}
}
\value{
get the image back as encoded file
}
\description{
creates a json output from the inputs
}
