% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{create_single_image_request}
\alias{create_single_image_request}
\title{helper function to create a list of details of one image annotation request}
\usage{
create_single_image_request(imagePath, feature, maxNumResults)
}
\arguments{
\item{imagePath}{character, file path, URL or Cloud Storage URI of the image}

\item{feature}{character, one out of: "LABEL_DETECTION", "FACE_DETECTION",
"TEXT_DETECTION", "DOCUMENT_TEXT_DETECTION", "LOGO_DETECTION",
"LANDMARK_DETECTION", "IMAGE_PROPERTIES", "OBJECT_LOCALIZATION"}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
list of request details for one image
}
\description{
creates a list output from the inputs
}
