% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{document_text_detection_extractor}
\alias{document_text_detection_extractor}
\title{helper function code to extract API response into a data.table for given feature type}
\usage{
document_text_detection_extractor(response)
}
\arguments{
\item{response}{an element of the API response object}
}
\value{
a data.table
}
\description{
helper function code to extract API response into a data.table for given feature type
}
