% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{encode_image}
\alias{encode_image}
\title{helper function to base64 encode the image file}
\usage{
encode_image(imagePath)
}
\arguments{
\item{imagePath}{character, path to the image}
}
\value{
get the image back as encoded file
}
\description{
base64 encodes an image file
}
