% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{extract_annotations}
\alias{extract_annotations}
\title{helper function code to extract the annotations}
\usage{
extract_annotations(responses, imagePaths, feature_type)
}
\arguments{
\item{responses}{an API response object}

\item{imagePaths}{character, file paths, URLs or Cloud Storage URIs of the images,
can be a combination of all three}

\item{feature_type}{the type of annotation as called in the response object}
}
\value{
a data.table
}
\description{
a utility to extract features from the API response
}
