% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{extract_error}
\alias{extract_error}
\title{helper function code to extract error from API response into a data.table}
\usage{
extract_error(responses, imagePaths)
}
\arguments{
\item{responses}{an API response object}

\item{imagePaths}{character, file paths, URLs or Cloud Storage URIs of the images,
can be a combination of all three}
}
\value{
a data.table
}
\description{
helper function code to extract error from API response into a data.table
}
