% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{extract_response}
\alias{extract_response}
\title{helper function code to extract the response data.frame}
\usage{
extract_response(pp, feature)
}
\arguments{
\item{pp}{an API response object}

\item{feature}{the name of the feature to return}
}
\value{
a data frame
}
\description{
a utility to extract features from the API response
}
