% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{extractor}
\alias{extractor}
\title{helper function code to provide an extractor function for different feature types}
\usage{
extractor(feature)
}
\arguments{
\item{feature}{character, one out of: "LABEL_DETECTION", "FACE_DETECTION",
"TEXT_DETECTION", "DOCUMENT_TEXT_DETECTION", "LOGO_DETECTION", "LANDMARK_DETECTION"}
}
\value{
a function
}
\description{
a utility to provide functions to extract features from the API response
}
