% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_feature_types.R
\name{gcv_get_available_feature_types}
\alias{gcv_get_available_feature_types}
\title{helper function code to record available feature types}
\usage{
gcv_get_available_feature_types()
}
\value{
a list of available features names and their types (as returned by the API)
}
\description{
helper function code to record available feature types
}
\examples{
gcv_get_available_feature_types()

}
