% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_search_response.R
\name{gcv_get_product_search_raw_response}
\alias{gcv_get_product_search_raw_response}
\title{Get raw results from the Vision API Product Search}
\usage{
gcv_get_product_search_raw_response(imagePath, project_id, location_id,
  product_set_id, product_category, maxNumResults)
}
\arguments{
\item{imagePath}{character, file path, URLs or Cloud Storage URI of the image}

\item{project_id}{character, GCP project id}

\item{location_id}{character, GCP location id}

\item{product_set_id}{character, product set id for Product Search}

\item{product_category}{character, roduct category for Product Search}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
a list with product search results
}
\description{
Given an image, and search parameters, this functions calls the
  Google Cloud Vision API, and returns an unformatted response (JSON) from the API.
  If you are looking for a friendlier result, refer to the
  `gcv_get_product_search_response()` function
}
