% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{get_available_feature_types}
\alias{get_available_feature_types}
\title{helper function code to record available feature types}
\usage{
get_available_feature_types()
}
\value{
a list of available features names and their types (as returned by the API)
}
\description{
helper function code to record available feature types
}
\examples{
get_available_feature_types()

}
