% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{get_feature_types}
\alias{get_feature_types}
\title{helper function code to record available feature types}
\usage{
get_feature_types()
}
\value{
a list of available features and their types
}
\description{
helper function code to record available feature types
}
