% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validations.R
\name{get_invalid_image_paths}
\alias{get_invalid_image_paths}
\title{helper function to validate input image paths}
\usage{
get_invalid_image_paths(vec)
}
\arguments{
\item{vec}{a vector of paths}
}
\value{
vector of invalid paths from @vec
}
\description{
helper function to validate input image paths
}
