% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{getBoundingBoxes}
\alias{getBoundingBoxes}
\title{helper function code to extract Bounding Box x,y coordinates for an API response element}
\usage{
getBoundingBoxes(response)
}
\arguments{
\item{response}{an element of the API response object}
}
\value{
a data.table
}
\description{
helper function code to extract Bounding Box x,y coordinates for an API response element
}
