% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{getGoogleVisionResponse}
\alias{getGoogleVisionResponse}
\title{Calling Google's Cloud Vision API}
\usage{
getGoogleVisionResponse(imagePath, feature = "LABEL_DETECTION",
  numResults = 5)
}
\arguments{
\item{imagePath}{path or url to the image}

\item{feature}{one out of: FACE_DETECTION, LANDMARK_DETECTION, LOGO_DETECTION, LABEL_DETECTION, TEXT_DETECTION}

\item{numResults}{the number of results to return.}
}
\value{
a data frame with results
}
\description{
input an image, provide the feature type and maxNumber of responses
}
\examples{
\dontrun{
    f <- system.file("extdata", "brandlogos.png", package = "RoogleVision")
    getGoogleVisionResponse(imagePath = f, feature = "LOGO_DETECTION")
}
}
