% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{image_to_text}
\alias{image_to_text}
\title{helper function to base64 encode the image file}
\usage{
image_to_text(imagePath)
}
\arguments{
\item{imagePath}{string, path or url to the image}
}
\value{
get the image back as encoded file
}
\description{
base64 encodes an image file
}
