% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{split_to_chunks}
\alias{split_to_chunks}
\title{helper function to split a vector to approximately equally sized chunks}
\usage{
split_to_chunks(vec, chunkSize)
}
\arguments{
\item{vec}{a vector}

\item{chunkSize}{integer, how long should the chunks be?}
}
\value{
a list of chunks
}
\description{
helper function to split a vector to approximately equally sized chunks
}
