% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{create_request_body}
\alias{create_request_body}
\title{helper function to create json for response request}
\usage{
create_request_body(encodedImages, feature, maxNumResults)
}
\arguments{
\item{encodedImages}{character, elements are outputs of image_to_text()}

\item{feature}{string, one out of: "FACE_DETECTION", "LANDMARK_DETECTION",
"LOGO_DETECTION", "LABEL_DETECTION", "TEXT_DETECTION"}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
request body (payload), encoded as json
}
\description{
creates a json output from the inputs
}
