% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{create_single_image_request}
\alias{create_single_image_request}
\title{helper function to create a list of details of one image annotation request}
\usage{
create_single_image_request(encodedImages, feature, maxNumResults)
}
\arguments{
\item{encodedImages}{character, elements are outputs of image_to_text()}

\item{feature}{string, one out of: "FACE_DETECTION", "LANDMARK_DETECTION",
"LOGO_DETECTION", "LABEL_DETECTION", "TEXT_DETECTION"}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
list of request details for one image
}
\description{
creates a list output from the inputs
}
