% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{extract_response}
\alias{extract_response}
\title{helper function code to extract the response data.frame}
\usage{
extract_response(responses, imagePaths, feature)
}
\arguments{
\item{responses}{an API response object}

\item{imagePaths}{character, file paths, URLs or Cloud Storage URIs of the images,
can be a combination of all three}

\item{feature}{character, one out of: "FACE_DETECTION", "LANDMARK_DETECTION",
"LOGO_DETECTION", "LABEL_DETECTION", "TEXT_DETECTION"}
}
\value{
a data frame
}
\description{
a utility to extract features from the API response
}
