% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{extractor}
\alias{extractor}
\title{helper function code to provide an extractor function for different feature types}
\usage{
extractor(feature_type)
}
\arguments{
\item{feature_type}{the type of annotation as called in the response object}
}
\value{
a function
}
\description{
a utility to provide functions to extract features from the API response
}
