% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{gcv_get_response}
\alias{gcv_get_response}
\title{Get parsed image annotations from the Google Cloud Vision API}
\usage{
gcv_get_response(imagePaths, feature, maxNumResults)
}
\arguments{
\item{imagePaths}{character, file paths, URLs or Cloud Storage URIs of the images,
can be a combination of all three}

\item{feature}{character, one out of: "LABEL_DETECTION", "FACE_DETECTION",
"TEXT_DETECTION", "DOCUMENT_TEXT_DETECTION", "LOGO_DETECTION", "LANDMARK_DETECTION"}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
a data frame with image annotation results
}
\description{
Given a list of images, a feature type and the maximum number of responses,
  this functions calls the Google Cloud Vision API, and returns the image annotations in a data.table format.
}
\examples{
\dontrun{
    imagePath <- system.file(
      "extdata", "golden_retriever_puppies.jpg", package = "googleCloudVisionR"
    )
    gcv_get_response(imagePaths = imagePath, maxNumResults = 7)
}

}
