% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{gcv_get_response}
\alias{gcv_get_response}
\title{helper function to call the API for one batch of images}
\usage{
gcv_get_response(imagePaths, feature, maxNumResults)
}
\arguments{
\item{imagePaths}{character, file paths, URLs or Cloud Storage URIs of the images,
can be a combination of all three}

\item{feature}{character, one out of: "FACE_DETECTION", "LABEL_DETECTION",
"TEXT_DETECTION", "DOCUMENT_TEXT_DETECTION", "LOGO_DETECTION"}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
a data frame with image annotation results
}
\description{
helper function to call the API for one batch of images
}
