% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{gcv_get_response}
\alias{gcv_get_response}
\title{Calling Google's Cloud Vision API}
\usage{
gcv_get_response(imagePaths, feature = "LABEL_DETECTION",
  maxNumResults = NULL)
}
\arguments{
\item{imagePaths}{string, paths or urls to the images}

\item{feature}{string, one out of: "FACE_DETECTION", "LANDMARK_DETECTION",
"LOGO_DETECTION", "LABEL_DETECTION", "TEXT_DETECTION"}

\item{maxNumResults}{integer, the maximum number of results to return.}
}
\value{
a data frame with results
}
\description{
input an image, provide the feature type and maxNumResults of responses
}
\examples{
\dontrun{
    f <- system.file("extdata", "brandlogos.png", package = "googleCloudVisionR")
    gcv_get_response(imagePaths = f, feature = "LOGO_DETECTION")
}

}
