% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{gcv_get_response}
\alias{gcv_get_response}
\title{helper function to call the API for one batch of images}
\usage{
gcv_get_response(imagePaths, feature, maxNumResults)
}
\arguments{
\item{imagePaths}{string, paths or urls to the images}

\item{feature}{string, one out of: "FACE_DETECTION", "LANDMARK_DETECTION",
"LOGO_DETECTION", "LABEL_DETECTION", "TEXT_DETECTION"}

\item{maxNumResults}{integer, the maximum number of results (per image) to be returned.}
}
\value{
a data frame with image annotation results
}
\description{
helper function to call the API for one batch of images
}
