% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudVisionR-package.R
\docType{package}
\name{googleCloudVisionR-package}
\alias{googleCloudVisionR}
\alias{googleCloudVisionR-package}
\title{googleCloudVisionR: Access to Google's Cloud Vision API for Image Recognition, OCR and
    Labeling}
\description{
This is a wrapper for Google's Cloud Vision API.
}
\author{
\strong{Maintainer}: Jeno Pal \email{paljenczy@gmail.com}

Authors:
\itemize{
  \item Tamas Koncz \email{t.koncz@gmail.com}
  \item Balazs Varkoly \email{varkoly.balazs@gmail.com}
  \item Eszter Kocsis \email{p.kocsis.eszter@gmail.com}
}

}
\keyword{internal}
