% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudVisionR-package.R
\docType{package}
\name{googleCloudVisionR-package}
\alias{googleCloudVisionR}
\alias{googleCloudVisionR-package}
\title{googleCloudVisionR: Access to 'Google Cloud Vision' API for Image Recognition, OCR and
    Labeling}
\description{
Interact with 'Google Cloud Vision' <https://cloud.google.com/vision/>
  API in R. Part of the 'cloudyr' <https://cloudyr.github.io/> project.
}
\author{
\strong{Maintainer}: Jeno Pal \email{paljenczy@gmail.com}

Authors:
\itemize{
  \item Tamas Koncz \email{t.koncz@gmail.com}
  \item Balazs Varkoly \email{varkoly.balazs@gmail.com}
  \item Eszter Kocsis \email{p.kocsis.eszter@gmail.com}
}

Other contributors:
\itemize{
  \item Florian Teschner [contributor]
}

}
\keyword{internal}
