% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_vision_response.R
\name{image_to_text}
\alias{image_to_text}
\title{helper function base_encode code the image file}
\usage{
image_to_text(imagePath)
}
\arguments{
\item{imagePath}{provide path/url to image}
}
\value{
get the image back as encoded file
}
\description{
base64 encodes an image file
}
